requirements.txt
# Core web framework
Flask>=2.0.0,<3.0.0

# Twilio for voice services
twilio>=8.0.0,<9.0.0

# Google Gemini AI integration
google-generativeai>=0.3.0

# Environment variable management
python-dotenv>=1.0.0

# HTTP requests (for integration with backend API)
requests>=2.28.0

# CORS handling for Flask
flask-cors>=3.0.10

# Production server
gunicorn>=20.1.0

# Database integration (if needed)
pymongo>=4.3.0

# For deployment monitoring
newrelic>=8.8.0,<9.0.0

